// 25032026/sce-front-auditoria/sce-front-auditoria/src/app/model/reportes/filtroReporteMesasObservaciones.ts
export class FiltroReporteMesasObservaciones {
    esquema: string;
    idProceso: number;
    idEleccion: number;
    eleccion: string;
    idCentroComputo: number;
    codigoCentroComputo: string;
    ubigeo: string;
    proceso: string;
    usuario: string;
    departamento?: string;
    provincia?: string;
    distrito?: string;
}